requirements.txt
beautifulsoup4==4.9.1
django-filter==2.3.0
django-heroku==0.3.1
djangorestframework==3.11.1
pyTelegramBotAPI==3.7.3
python-telegram-bot-pagination==0.0.1
requests==2.23.0
gunicorn
psycopg2
mysqlclient
